Paging list of sets in enterprise server.

+++ demeter-core/src/main/java/com/panforge/demeter/core/content/PageCursor.java
@@ -23,5 +23,5 @@ public interface PageCursor {
    * Gets cursor position.
    * @return cursor position
    */
-  int cursor();
+  long cursor();
 }
